requirements.txt
anndata==0.12.3
gensim==4.4.0
h5py==3.15.1
lifelines==0.30.0
matplotlib==3.10.7
numpy==2.3.4
pandas==2.3.3
Pillow==12.0.0
pytorch_lightning==2.5.5
scikit_learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
Shapely==2.1.2
statsmodels==0.14.5
torch==2.9.0
torch_geometric==2.7.0
torchmetrics==1.8.2
torchvision==0.24.0
tqdm==4.67.1
trident==1.4.2
